+++ evaluator/src/commonTest/kotlin/com/github/jcornaz/kwik/evaluator/CheckForAll4Test.kt
@@ -1,7 +1,6 @@
 package com.github.jcornaz.kwik.evaluator
 
 import com.github.jcornaz.kwik.generator.api.Generator
-import com.github.jcornaz.kwik.generator.api.withSamples
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
@@ -25,34 +24,6 @@ class CheckForAll4Test : AbstractRunnerTest() {
         ) { _, _, _, _ ->
             assertTrue(invocation())
         }
-    }
-
-    @Test
-    fun evaluateSamples() {
-        val ints = mutableSetOf<Int>()
-        val doubles = mutableSetOf<Double>()
-        val longs = mutableSetOf<Long>()
-        val floats = mutableSetOf<Float>()
-
-        val gen1 = Generator.create { it.nextInt(0, 10) }.withSamples(42, 100)
-        val gen2 = Generator.create { it.nextDouble(0.0, 10.0) }.withSamples(123.0, 678.0)
-        val gen3 = Generator.create { it.nextLong(0L, 10L) }.withSamples(-42L)
-        val gen4 = Generator.create { it.nextDouble(0.0, 10.0).toFloat() }.withSamples(-6f, 18f)
-
-        checkForAll(gen1, gen2, gen3, gen4) { i, d, l, f ->
-            ints += i
-            doubles += d
-            longs += l
-            floats += f
-        }
-
-        assertTrue(42 in ints)
-        assertTrue(100 in ints)
-        assertTrue(123.0 in doubles)
-        assertTrue(123.0 in doubles)
-        assertTrue(-42L in longs)
-        assertTrue(-6f in floats)
-        assertTrue(18f in floats)
     }
 
     @Test