+++ generator/api/src/commonMain/kotlin/com/github/jcornaz/kwik/generator/api/GeneratorOperators.kt
@@ -2,6 +2,8 @@ package com.github.jcornaz.kwik.generator.api
 
 import kotlin.random.Random
 
+private const val DEFAULT_SAMPLE_PROBABILITY = 0.2
+
 /**
  * Returns a generator containing the results of applying the given transform function to each element emitted by
  * the original generator.
@@ -11,7 +13,7 @@ fun <T, R> Generator<T>.map(transform: (T) -> R): Generator<R> =
 
 private class MapGenerator<T, R>(private val source: Generator<T>, private val transform: (T) -> R) :
     Generator<R> {
-    override val samples: Set<R> = source.samples.mapTo(mutableSetOf(), transform)
+    override val samples: Set<R> get() = source.samples.mapTo(mutableSetOf(), transform)
 
     override fun generate(random: Random): R = transform(source.generate(random))
 }
@@ -36,8 +38,8 @@ fun <T, R> Generator<T>.andThen(transform: (T) -> Generator<R>): Generator<R> =
 fun <T, R> Generator<T>.flatMap(transform: (T) -> Generator<R>): Generator<R> = andThen(transform)
 
 private class AndThenGenerator<T, R>(private val source: Generator<T>, private val transform: (T) -> Generator<R>): Generator<R> {
-    override val samples: Set<R> =
-        source.samples.flatMapTo(HashSet()) { transform(it).samples }
+    override val samples: Set<R>
+        get() = source.samples.flatMapTo(HashSet()) { transform(it).samples }
 
     override fun generate(random: Random): R =
         source.generate(random).let(transform).generate(random)
@@ -65,7 +67,7 @@ private class FilterGenerator<T>(
     private val source: Generator<T>,
     private val predicate: (T) -> Boolean
 ) : Generator<T> {
-    override val samples: Set<T> = source.samples.filterTo(mutableSetOf(), predicate)
+    override val samples: Set<T> get() = source.samples.filterTo(mutableSetOf(), predicate)
 
     override fun generate(random: Random): T {
         var value = source.generate(random)
@@ -78,23 +80,33 @@ private class FilterGenerator<T>(
 }
 
 /**
- * Returns a new generator adding the given [samples] into generated random values.
- *
- * The "random" values always start by the given [samples] so that they always appear at least once.
+ * Returns a new generator that has a good [probability] to generate a value from the given [samples],
+ * and generate from source the rest of the time.
  */
-fun <T> Generator<T>.withSamples(vararg samples: T): Generator<T> =
-    withSamples(samples.asList())
+fun <T> Generator<T>.withSamples(vararg samples: T, probability: Double = DEFAULT_SAMPLE_PROBABILITY): Generator<T> =
+    withSamples(samples.asList(), probability)
 
 /**
- * Returns a new generator adding the given [samples] into generated random values.
- *
- * The "random" values always start by the given [samples] so that they always appear at least once.
+ * Returns a new generator that has a good [probability] to generate a value from the given [samples],
+ * and generate from source the rest of the time.
  */
-fun <T> Generator<T>.withSamples(samples: Iterable<T>): Generator<T> {
+fun <T> Generator<T>.withSamples(samples: Iterable<T>, probability: Double = DEFAULT_SAMPLE_PROBABILITY): Generator<T> {
+    require(probability > 0.0 && probability < 1.0) {
+        "Invalid sample probability: $probability. Must be greater than 0 and smaller than 1"
+    }
+
     val sampleList = (samples as? List<T>) ?: samples.toList()
     if (sampleList.isEmpty()) return this
 
-    return SampleGenerator(this, sampleList)
+    val frequencyGen = Generator.frequency(
+        probability to Generator.of(samples),
+        (1 - probability) to this
+    )
+
+    return object : Generator<T> by frequencyGen {
+        override val samples: Set<T>
+            get() = samples.toSet()
+    }
 }
 
 /**
@@ -105,20 +117,7 @@ fun <T> Generator<T>.withSamples(samples: Iterable<T>): Generator<T> {
 fun <T> Generator<T>.withNull(): Generator<T?> =
     NullGenerator(this)
 
-private class SampleGenerator<T>(
-    private val source: Generator<T>,
-    samples: Iterable<T>
-) : Generator<T> by source {
-
-    override val samples: Set<T> = source.samples + samples
-
-    init {
-        require(this.samples.isNotEmpty()) { "No sample provided" }
-    }
-}
-
 private class NullGenerator<T>(private val source: Generator<T>) : Generator<T?> {
-
     override val samples: Set<T?> = source.samples.plus<T?>(null)
 
     override fun generate(random: Random): T? = source.generate(random)