+++ CHANGELOG.rst
@@ -14,6 +14,13 @@ and this project adheres to `Semantic Versioning`_.
 Unreleased_
 -----------
 
+Changed
+.......
+
+All generators no longer statically provide ``samples``. Instead they have a good probability to generate edge cases.
+
+* ``withSample`` now takes a probability argument, and include the samples in the generation
+
 Added
 .....